Created a seperate page handling file.
